+++ src/__tests__/mocks/RPClientMock.ts
@@ -1,6 +1,7 @@
 import { ReportPortalConfig } from '../../models';
 
-export const mockedDate = '2024-09-23T12:20:59.392987Z';
+export const mockedDate = Date.now();
+
 export class RPClientMock {
   private config: ReportPortalConfig;
 
@@ -28,9 +29,21 @@ export class RPClientMock {
 
   public sendLog = jest.fn().mockReturnValue({
     promise: Promise.resolve('ok'),
+    tempId: 'tempLogId',
   });
 
   public checkConnect = jest.fn().mockReturnValue({
     promise: Promise.resolve('ok'),
   });
+
+  public updateLaunch = jest.fn().mockReturnValue({
+    promise: Promise.resolve('ok'),
+    tempId: 'tempLaunchId',
+  });
+
+  public getPromiseFinishAllItems = jest.fn().mockResolvedValue('ok');
+
+  public helpers = {
+    now: jest.fn().mockReturnValue(Date.now()),
+  };
 }