+++ src/constants/index.ts
@@ -18,5 +18,5 @@
 export { LAUNCH_MODES } from './launchModes';
 export { TEST_ITEM_TYPES } from './testItemTypes';
 export { STATUSES } from './statuses';
-export { LOG_LEVELS } from './logLevels';
+export { PREDEFINED_LOG_LEVELS, LOG_LEVELS } from './logLevels';
 export { TASK_STATUS, TASK_MODE, FINISHED_STATES } from './vitest';